backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=9.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-dateutil>=2.8.0
pytz>=2023.3
django-cors-headers>=4.0.0
django-unfold>=0.15.0
dj-database-url>=2.0.0
python-decouple>=3.8
requests>=2.31.0
